requirements.txt
Flask==2.0.2
PyJWT==1.7.1
flask_sqlalchemy
gunicorn